@@ -1,6 +1,4 @@
-﻿using System;
-using System.Linq;
-using System.Linq.Expressions;
+﻿using System.Linq;
 using Serilog.Sanitizer.Extensions;
 using Serilog.Sanitizer.UnitTests.Sinks;
 using Serilog.Events;
@@ -9,8 +7,10 @@ using NUnit.Framework;
 namespace Serilog.Sanitizer.UnitTests
 {
     [TestFixture]
-    public class SanitizerTests
+    public class StructuredDataTests
     {
+        private const string OverrideValue = "***** VALUE OVERRIDDEN *****";
+
         internal class Employee
         {
             public string FirstName { get; set; }
@@ -18,54 +18,15 @@ namespace Serilog.Sanitizer.UnitTests
             public string Ssn { get; set; }
         }
 
-        private const string OverrideValue = "***** VALUE OVERRIDDEN *****";
-
-        [Test]
-        public void UnstructuredValueIsOverriddenTuple()
-        {
-            LogEvent evt = null;
-
-            var logger = new LoggerConfiguration()
-                            .Sanitize()
-                                .Unstructured()
-                                .ByOverriding(("test", OverrideValue))
-                                .Continue()
-                            .WriteTo.Sink(new DelegatingSink(e => evt = e))
-                            .CreateLogger();
-
-            logger.Information("Sensitive Information {@test}", "original value");
-
-            Assert.AreEqual(OverrideValue, ((ScalarValue)evt.Properties["test"]).Value);
-        }
-
-        [Test]
-        public void UnstructuredValueIsOverriddenNonTuple()
-        {
-            LogEvent evt = null;
-
-            var logger = new LoggerConfiguration()
-                            .Sanitize()
-                                .Unstructured()
-                                .ByOverriding("test", OverrideValue)
-                                .Continue()
-                            .WriteTo.Sink(new DelegatingSink(e => evt = e))
-                            .CreateLogger();
-
-            logger.Information("Sensitive Information {@test}", "original value");
-
-            Assert.AreEqual(OverrideValue, ((ScalarValue)evt.Properties["test"]).Value);
-        }
-
         [Test]
         public void StructuredValueIsRemoved()
         {
             LogEvent evt = null;
-            Expression<Func<Employee, object>> valueTypeProperty = dm => dm.Ssn;
 
             var logger = new LoggerConfiguration()
                             .Sanitize()
                                 .Structured()
-                                    .ByRemoving(valueTypeProperty)
+                                    .ByRemoving<Employee>(x => x.Ssn)
                                 .Continue()
                             .WriteTo.Sink(new DelegatingSink(e => evt = e))
                             .CreateLogger();
@@ -80,12 +41,11 @@ namespace Serilog.Sanitizer.UnitTests
         public void StructuredValueIsOverridden()
         {
             LogEvent evt = null;
-            (Expression<Func<Employee, object>>, string) toOverride = (dm => dm.Ssn, OverrideValue);
 
             var logger = new LoggerConfiguration()
                             .Sanitize()
                                 .Structured()
-                                    .ByOverriding<Employee>(toOverride)
+                                    .ByOverriding<Employee>((x => x.Ssn, OverrideValue))
                                 .Continue()
                             .WriteTo.Sink(new DelegatingSink(e => evt = e))
                             .CreateLogger();
